Square the tone plot, and put it with the other widgets
The plot was never square. Its correction could only add height, but the
tick labels are wider at the left than they are tall at the bottom, so
squaring it means taking nine pixels away; the clamp at zero made sure it
never could. Measured, it sat at 176x185 every frame with the correction
pinned at 0. Letting it go negative settles at 176x176 after one frame.

ToneCurvePlot lives in imgui_ext now, where the other widgets are. Its
interface only ever mentions ImVec4 and float2, so the declaration goes in
the header and the bodies in the .cpp, and implot stays out of a header
almost everything includes.

Three smaller things: alpha_channel_index listed the same four group types
group_has_alpha does, and now asks it; subject_channels and
subject_color_groups sit in subject.cpp beside their declarations, taking
their file-local helper with them; and a comment still placed the CIE
diagram in image-histogram.cpp, which it left in the split.

300 cases, 361,742 assertions; hdrview_gui_tests 89/89.
