+++ CMakeLists.txt
@@ -1730,7 +1730,7 @@ set(HDRVIEW_SOURCES
     src/edit/envmap.cpp
     src/edit/poisson.cpp
     src/edit/filters.cpp
-    src/edit/edit_ops.cpp
+    src/edit/edit_ops.cpp src/edit/subject.cpp
     src/edit/undo.cpp
     src/dithermatrix256.cpp
     src/platform_utils.cpp
