+++ src/edit/edit_ops.cpp
@@ -18,22 +18,6 @@ using std::string;
 namespace
 {
 
-/// The groups \p subject's scope names, before any filtering by what they contain.
-std::vector<int> subject_groups(const Image &img, const EditSubject &subject)
-{
-    std::vector<int> groups;
-    if (subject.scope == EditSubject::Scope_AllChannels)
-    {
-        for (int g = 0; g < int(img.groups.size()); ++g) groups.push_back(g);
-    }
-    else if (subject.scope == EditSubject::Scope_SelectedGroups)
-        groups = img.selected_groups();
-    else if (int g = img.active_group_index(Target_Primary); img.is_valid_group(g))
-        groups.push_back(g);
-
-    return groups;
-}
-
 /// Compute a rectangle of one or more channels in parallel, then upload it.
 /**
     \p op fills the staging arrays for the rows [y0, y1) of the rectangle: one array per entry of
@@ -79,48 +63,6 @@ std::vector<int> target_groups(const ConstImagePtr &img, int pointed_at)
         groups.push_back(img->active_group_index(Target_Primary));
 
     return groups;
-}
-
-std::vector<int> subject_channels(const Image &img, const EditSubject &subject)
-{
-    // every channel, not the union of every group's, so one belonging to no group is still covered
-    if (subject.scope == EditSubject::Scope_AllChannels)
-    {
-        std::vector<int> channels(img.channels.size());
-        std::iota(channels.begin(), channels.end(), 0);
-        return channels;
-    }
-
-    std::vector<int> channels;
-    for (int g : subject_groups(img, subject))
-    {
-        const auto &group = img.groups[size_t(g)];
-        for (int c = 0; c < group.num_channels; ++c) channels.push_back(group.channels[c]);
-    }
-    return channels;
-}
-
-std::pair<std::vector<int>, std::vector<int>> subject_color_groups(const Image &img, const EditSubject &subject)
-{
-    std::vector<int> groups = subject_groups(img, subject);
-
-    groups.erase(std::remove_if(groups.begin(), groups.end(),
-                                [&img](int g)
-                                {
-                                    const auto t = img.groups[size_t(g)].type;
-                                    return t != ChannelGroup::RGB_Channels && t != ChannelGroup::RGBA_Channels;
-                                }),
-                 groups.end());
-
-    // every channel of every covered group, which is the set an undo entry has to hold
-    std::vector<int> channels;
-    for (int g : groups)
-    {
-        const auto &group = img.groups[size_t(g)];
-        for (int c = 0; c < group.num_channels; ++c) channels.push_back(group.channels[c]);
-    }
-
-    return {groups, channels};
 }
 
 std::pair<std::vector<int>, Box2i> resolve_subject(const ConstImagePtr &img, const EditSubject &subject,
